+++ retrievers.h
@@ -21,7 +21,7 @@ namespace retrievers
             : deal_type(other.deal_type), start_seat(move(other.start_seat)), cards(move(other.cards)) {}
         ~deal_result() = default;
 
-        int8_t deal_type;
+        int16_t deal_type;
         std::string start_seat;
         std::vector<std::string> cards;
     };
@@ -33,7 +33,7 @@ namespace retrievers
             : trick_type(other.trick_type), cards(move(other.cards)) {}
         ~trick_result() = default;
 
-        int8_t trick_type;
+        int16_t trick_type;
         std::vector<std::string> cards;
     };
 
@@ -44,7 +44,7 @@ namespace retrievers
             : trick_type(other.trick_type), cards(move(other.cards)), taking_seat(move(other.taking_seat)) {}
         ~taken_result() = default;
 
-        int8_t trick_type;
+        int16_t trick_type;
         std::vector<std::string> cards;
         std::string taking_seat;
     };
@@ -53,7 +53,7 @@ namespace retrievers
     string retrieve_busy(const string& message);
     deal_result retrieve_deal(const string& message);
     trick_result retrieve_trick(const string& message);
-    int8_t retrieve_wrong(const string& message);
+    int16_t retrieve_wrong(const string& message);
     taken_result retrieve_taken(const string& message);
     map<string, int32_t> retrieve_score(const string& message);
     map<string, int32_t> retrieve_total(const string& message);
