+++ senders.cpp
@@ -12,7 +12,7 @@ ssize_t senders::send_busy(int socket_fd, const string& seats)
     return common::write_to_socket(socket_fd, message.data(), message.length());
 }
 
-ssize_t senders::send_deal(int socket_fd, int8_t deal_type, const string& start_seat, const vector<string>& cards)
+ssize_t senders::send_deal(int socket_fd, int16_t deal_type, const string& start_seat, const vector<string>& cards)
 {
     string message = "DEAL " + std::to_string(deal_type) + start_seat;
     for (const string& card : cards) { message += card; }
@@ -20,7 +20,7 @@ ssize_t senders::send_deal(int socket_fd, int8_t deal_type, const string& start_
     return common::write_to_socket(socket_fd, message.data(), message.length());
 }
 
-ssize_t senders::send_trick(int socket_fd, int8_t trick_type, const vector<string>& cards)
+ssize_t senders::send_trick(int socket_fd, int16_t trick_type, const vector<string>& cards)
 {
     string message = "TRICK " + std::to_string(trick_type);
     for (const string& card : cards) { message += card; }
@@ -28,13 +28,13 @@ ssize_t senders::send_trick(int socket_fd, int8_t trick_type, const vector<strin
     return common::write_to_socket(socket_fd, message.data(), message.length());
 }
 
-ssize_t senders::send_wrong(int socket_fd, int8_t trick_type)
+ssize_t senders::send_wrong(int socket_fd, int16_t trick_type)
 {
     string message = "WRONG " + std::to_string(trick_type) + DELIMETER;
     return common::write_to_socket(socket_fd, message.data(), message.length());
 }
 
-ssize_t senders::send_taken(int socket_fd, int8_t trick_type, const vector<string>& cards, const string& taking_seat)
+ssize_t senders::send_taken(int socket_fd, int16_t trick_type, const vector<string>& cards, const string& taking_seat)
 {
     string message = "TAKEN " + std::to_string(trick_type);
     for (const string& card : cards) { message += card; }
